% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bencalibr.R
\name{bencalibr}
\alias{bencalibr}
\title{Calibration for benefit plots}
\usage{
bencalibr(
  data = NULL,
  Ngroups = 5,
  y.observed,
  treat,
  predicted.treat.0,
  predicted.treat.1,
  type = "continuous",
  smoothing.function = "rlm",
  measure = "RD"
)
}
\arguments{
\item{data}{A dataframe}
}
\value{
Aplot
}
\description{
Produces plot
}
\examples{
temp1 <- F_to_C(50);
temp2 <- F_to_C( c(50, 63, 23) );
}
