% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_binary3.R
\name{expit}
\alias{expit}
\title{Expit}
\usage{
expit(x)
}
\arguments{
\item{x}{A real number}
}
\value{
exp(x)/(1+exp(x))
}
\description{
Calculates the expit of a real number
}
\examples{
expit(2.3)
}
