% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predcompare.R
\name{predcompare}
\alias{predcompare}
\title{Calibraton measures for benefit}
\usage{
predcompare(
  repeats = 50,
  Ngroups = 10,
  X,
  treat,
  Y,
  predicted.treat.1,
  predicted.treat.0,
  type = "continuous",
  measure = "RD"
)
}
\arguments{
\item{repeats}{A dataframe}
}
\value{
A table
}
\description{
Calculates measures
}
\examples{
temp1 <- F_to_C(50);
temp2 <- F_to_C( c(50, 63, 23) );
}
