% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_binary.R
\name{simbinary}
\alias{simbinary}
\title{Simulate binary}
\usage{
simbinary(Npat = 100)
}
\arguments{
\item{repeats}{A dataframe}
}
\value{
A table
}
\description{
Calculates measures
}
\examples{
temp1 <- F_to_C(50);
temp2 <- F_to_C( c(50, 63, 23) );
}
