% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_continuous.R
\name{simulate.data.cont}
\alias{simulate.data.cont}
\title{Simulate continuous}
\usage{
\method{simulate}{data.cont}(Npat)
}
\arguments{
\item{repeats}{A dataframe}
}
\value{
A table
}
\description{
Calculates measures
}
\examples{
temp1 <- F_to_C(50);
temp2 <- F_to_C( c(50, 63, 23) );
}
